refactor: Move dangling references removal to select builder function
